.history/backend/requirements_20250829172125.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
